feat(node/util): Move inspect fnction to internal (TS -> JS)
